Typecast escaped values for multiselect fields
The 2.2.5.3 CMB2 update causes the function for getting and sorting the stored multiselect options to fail when only a single value is retrieved from the db - returning a string instead of the expected array. Typecasting in the helper function instead of the filter function might not be a proper solution, but it seems to work.